# specs/006-constellation-stack/contracts/control-plane.openapi.yaml
openapi: 3.1.0
info:
  title: Constellation Control Plane
  version: 0.1.0
  description: Internal authenticated contract between Nova Cloud, Habitat runtime control, and Horizon.
servers:
  - url: https://control.internal
security:
  - bearerAuth: []
paths:
...
  /v1/studios/{studioId}/domains:
    post:
      operationId: bindDomain
      parameters:
        - $ref: "#/components/parameters/StudioId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host, kind, deploymentId]
              properties:
                host: { type: string, format: hostname }
                kind: { type: string, enum: [platform, custom] }
                deploymentId: { type: string }
      responses:
        "201": { description: Domain binding created }
        "409": { description: Host is owned by another tenant }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    StudioId:
      name: studioId
      in: path
      required: true
      schema: { type: string }
...